[refact] join api 분리
